refactor(ui): move SystemActions from utils/ to platform/
Second step in utils/ breakup. SystemActions is purely OS-seam code
(Desktop.getDesktop().open / browse), separate concern from
identity (skin) and console (game-output buffer) it used to sit
next to. New platform/ package becomes the home for OS-level
wrappers; future additions like clipboard / system tray / file-
picker helpers land here.

No behaviour change; package declaration and seven call-site
imports updated.
