+++ client-ui/src/desktopMain/kotlin/hivens/ui/screens/ServerSettingsScreen.kt
@@ -54,7 +54,7 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.koin.compose.koinInject
-import hivens.ui.utils.SystemActions
+import hivens.ui.platform.SystemActions
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
